# chore: improve show streams performance