--- a/tests/suites/5_ee/05_stream/05_0000_ee_stream.result
+++ b/tests/suites/5_ee/05_stream/05_0000_ee_stream.result
@@ -1,7 +1,7 @@
 true
 2	INSERT	false	true
-test_s default default db_stream.t
-test_s1 default default db_stream.t
+test_s	db_stream.t	append_only
+test_s1	db_stream.t	standard
 test_s default default db_stream.t NULL test append_only
 test_s1 default default db_stream.t NULL standard standard
 test_s default default db_stream.t NULL test append_only