--- a/tests/suites/5_ee/05_stream/05_0000_ee_stream.sh
+++ b/tests/suites/5_ee/05_stream/05_0000_ee_stream.sh
@@ -17,7 +17,7 @@ echo "optimize table db_stream.t compact" | $BENDSQL_CLIENT_CONNECT
 echo "select a, change\$action, change\$is_update, change\$row_id='$BASE_ROW_ID' from default.test_s" | $BENDSQL_CLIENT_CONNECT
 
 echo "create stream test_s1 on table db_stream.t at(stream => default.test_s) append_only=false comment = 'standard'" | $BENDSQL_CLIENT_CONNECT
-echo "show streams like 'test_s%'" | $BENDSQL_CLIENT_CONNECT | awk '{print $(NF-3), $(NF-2), $(NF-1), $NF}'
+echo "show streams like 'test_s%'" | $BENDSQL_CLIENT_CONNECT
 echo "show full streams like 'test_s%'" | $BENDSQL_CLIENT_CONNECT | awk '{print $(NF-6), $(NF-5), $(NF-4), $(NF-3), $(NF-2), $(NF-1), $NF}'
 echo "desc stream default.test_s" | $BENDSQL_CLIENT_CONNECT | awk '{print $(NF-6), $(NF-5), $(NF-4), $(NF-3), $(NF-2), $(NF-1), $NF}'