I hereby agree to the terms of the CLA available at: https://docs.databend.com/dev/policies/cla/

## Summary

Checking whether a stream is valid or not is very **time-consuming**.

1. Add the system table `streams_terse`, which is a terse stream information table. Used for `show streams`.
```sql
mysql> show columns in streams from system;
+-------------------+-----------------+------+---------+-------+------+
| Field             | Type            | Null | Default | Extra | Key  |
+-------------------+-----------------+------+---------+-------+------+
| catalog           | VARCHAR         | NO   |         |  NULL | NULL |
| comment           | VARCHAR         | NO   |         |  NULL | NULL |
| created_on        | TIMESTAMP       | NO   |         |  NULL | NULL |
| database          | VARCHAR         | NO   |         |  NULL | NULL |
| invalid_reason    | VARCHAR         | NO   |         |  NULL | NULL |
| mode              | VARCHAR         | NO   |         |  NULL | NULL |
| name              | VARCHAR         | NO   |         |  NULL | NULL |
| owner             | VARCHAR         | YES  |         |  NULL | NULL |
| snapshot_location | VARCHAR         | YES  |         |  NULL | NULL |
| stream_id         | BIGINT UNSIGNED | NO   |         |  NULL | NULL |
| table_id          | BIGINT UNSIGNED | NO   |         |  NULL | NULL |
| table_name        | VARCHAR         | NO   |         |  NULL | NULL |
| table_version     | BIGINT UNSIGNED | NO   |         |  NULL | NULL |
| updated_on        | TIMESTAMP       | NO   |         |  NULL | NULL |
+-------------------+-----------------+------+---------+-------+------+
14 rows in set (0.14 sec)
Read 14 rows, 1.62 KiB in 0.020 sec., 706.25 rows/sec., 81.48 KiB/sec.

mysql> show columns in streams_terse from system;
+------------+---------+------+---------+-------+------+
| Field      | Type    | Null | Default | Extra | Key  |
+------------+---------+------+---------+-------+------+
| catalog    | VARCHAR | NO   |         |  NULL | NULL |
| database   | VARCHAR | NO   |         |  NULL | NULL |
| mode       | VARCHAR | NO   |         |  NULL | NULL |
| name       | VARCHAR | NO   |         |  NULL | NULL |
| table_name | VARCHAR | NO   |         |  NULL | NULL |
+------------+---------+------+---------+-------+------+
5 rows in set (0.07 sec)
Read 5 rows, 635.00 B in 0.014 sec., 363.48 rows/sec., 45.08 KiB/sec.
```
2. Modify the display information of `show streams` to `name | table_on| mode `.
```sql
mysql> show streams;
+--------------------+-------------+-------------+
| Streams_in_default | table_on    | mode        |
+--------------------+-------------+-------------+
| s                  | default.t   | append_only |
| test_s             | db_stream.t | append_only |
| test_s1            | db_stream.t | standard    |
+--------------------+-------------+-------------+
3 rows in set (0.05 sec)
Read 3 rows, 277.00 B in 0.035 sec., 84.73 rows/sec., 7.64 KiB/sec.

mysql> show full streams;
+----------------------------+---------+----------+---------+-------------+-------------+-------+----------+-------------------------------------------------------------------------+
| created_on                 | name    | database | catalog | table_on    | mode        | owner | comment  | invalid_reason                                                          |
+----------------------------+---------+----------+---------+-------------+-------------+-------+----------+-------------------------------------------------------------------------+
| 2024-05-09 15:18:24.975373 | s       | default  | default | default.t   | append_only | NULL  |          | Base table 'default'.'t' dropped, cannot read from stream 'default'.'s' |
| 2024-05-09 16:33:42.259303 | test_s  | default  | default | db_stream.t | append_only | NULL  | test     |                                                                         |
| 2024-05-09 16:34:24.810181 | test_s1 | default  | default | db_stream.t | standard    | NULL  | standard |                                                                         |
+----------------------------+---------+----------+---------+-------------+-------------+-------+----------+-------------------------------------------------------------------------+
3 rows in set (0.07 sec)
Read 3 rows, 712.00 B in 0.046 sec., 65.55 rows/sec., 15.19 KiB/sec.
```
3. Parallelize the operation of checking the validity of streams in the system table streams.



The first run without cache(Rows: 204).  **Performance**:
ver| show streams | show full streams
-|- | -
`Before`|5.1s | 5.1s
`After`| 167ms | 1.1s

- Fixes #15452

## Tests

- [ ] Unit Test
- [ ] Logic Test
- [ ] Benchmark Test
- [x] No Test  - _Explain why_

## Type of change

- [ ] Bug Fix (non-breaking change which fixes an issue)
- [ ] New Feature (non-breaking change which adds functionality)
- [ ] Breaking Change (fix or feature that could cause existing functionality not to work as expected)
- [ ] Documentation Update
- [ ] Refactoring
- [x] Performance Improvement
- [ ] Other (please describe):

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/datafuselabs/databend/15460)
<!-- Reviewable:end -->